requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
azure-core==1.29.5
    # via
    #   azure-identity
    #   azure-storage-blob
azure-identity==1.15.0
    # via -r requirements.in
azure-storage-blob==12.19.0
    # via -r requirements.in
cachetools==5.3.0
    # via google-auth
certifi==2023.7.22
    # via
    #   elasticsearch
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
cryptography==41.0.2
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.20
    # via
    #   -r requirements.in
    #   django-anymail
    #   django-cors-headers
    #   django-csp
    #   django-guardian
    #   django-hijack
    #   django-storages
django-anymail==9.0
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-csp==3.7
    # via -r requirements.in
django-guardian==2.4.0
    # via -r requirements.in
django-hijack==3.2.6
    # via -r requirements.in
django-storages[google]==1.11.1
    # via
    #   -r requirements.in
    #   django-storages
elasticsearch==7.9.1
    # via
    #   -r requirements.in
    #   elasticsearch-dsl
elasticsearch-dsl==7.2.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
google-api-core==2.11.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.14.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.3.2
    # via google-cloud-storage
google-cloud-storage==2.7.0
    # via django-storages
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.1
    # via google-cloud-storage
googleapis-common-protos==1.58.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
isodate==0.6.1
    # via azure-storage-blob
jmespath==1.0.1
    # via -r requirements.in
msal==1.25.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
openpyxl==3.1.1
    # via -r requirements.in
pillow==9.4.0
    # via -r requirements.in
portalocker==2.8.2
    # via msal-extensions
protobuf==3.20.2
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2==2.9.5
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via
    #   msal
    #   social-auth-core
pyliftover==0.4
    # via -r requirements.in
python-dateutil==2.8.2
    # via elasticsearch-dsl
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.7.1
    # via django
redis==4.5.4
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   azure-core
    #   django-anymail
    #   google-api-core
    #   google-cloud-storage
    #   msal
    #   requests-oauthlib
    #   requests-toolbelt
    #   slacker
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
requests-toolbelt==0.10.1
    # via -r requirements.in
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   azure-core
    #   elasticsearch-dsl
    #   google-auth
    #   isodate
    #   python-dateutil
slacker==0.14.0
    # via -r requirements.in
slugify==0.0.1
    # via -r requirements.in
social-auth-app-django==5.0.0
    # via -r requirements.in
social-auth-core==4.3.0
    # via
    #   -r requirements.in
    #   social-auth-app-django
sqlparse==0.4.4
    # via django
tqdm==4.64.1
    # via -r requirements.in
typing-extensions==4.8.0
    # via
    #   azure-core
    #   azure-storage-blob
urllib3==1.26.14
    # via
    #   elasticsearch
    #   requests
whitenoise==6.3.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
